refactor: CommentRequest -> CommentCondition 으로 이름 변경
