+++ src/main/java/com/sprint/monew/domain/comment/repository/CommentRepositoryCustom.java
@@ -1,13 +1,13 @@
 package com.sprint.monew.domain.comment.repository;
 
 import com.sprint.monew.domain.comment.dto.CommentDto;
-import com.sprint.monew.domain.comment.dto.request.CommentRequest;
+import com.sprint.monew.domain.comment.dto.CommentCondition;
 import java.util.UUID;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 
 public interface CommentRepositoryCustom {
 
-  Slice<CommentDto> getComments(CommentRequest condition, UUID userId, Pageable pageable);
+  Slice<CommentDto> getComments(CommentCondition condition, UUID userId, Pageable pageable);
 
 }
